+++ selftest-core/src/main/java/net/oneandone/httpselftest/http/FullTestResponse.java
@@ -1,0 +1,13 @@
+package net.oneandone.httpselftest.http;
+
+import net.oneandone.httpselftest.test.api.TestResponse;
+
+public interface FullTestResponse {
+
+    TestResponse getTestResponse();
+
+    String headerBlock();
+
+    String bodyBlock();
+
+}