+++ selftest-core/src/main/java/net/oneandone/httpselftest/http/HttpClient.java
@@ -2,7 +2,7 @@ package net.oneandone.httpselftest.http;
 
 public interface HttpClient {
 
-    TestResponse call(String baseUrl, TestRequest request, String runId, int timeoutMillis);
+    FullTestResponse call(String baseUrl, TestRequest request, String runId, int timeoutMillis);
 
     enum Type {
         URLCON, SOCKET;