+++ selftest-core/src/main/java/net/oneandone/httpselftest/http/urlcon/UrlConnectionTestResponse.java
@@ -3,10 +3,10 @@ package net.oneandone.httpselftest.http.urlcon;
 import java.util.List;
 
 import net.oneandone.httpselftest.http.Headers;
-import net.oneandone.httpselftest.http.TestResponse;
-import net.oneandone.httpselftest.test.api.AssertionException;
+import net.oneandone.httpselftest.http.FullTestResponse;
+import net.oneandone.httpselftest.test.api.TestResponse;
 
-public class UrlConnectionTestResponse implements TestResponse {
+public class UrlConnectionTestResponse implements FullTestResponse {
 
     final int statusCode;
     final String statusLine;
@@ -27,40 +27,23 @@ public class UrlConnectionTestResponse implements TestResponse {
     }
 
     @Override
-    public int getStatus() {
-        return statusCode;
+    public TestResponse getTestResponse() {
+        return new TestResponse(statusCode, headers, body);
     }
 
     @Override
-    public String getHeader(String name) {
-        List<String> values = headers.get(name);
-        if (values == null) {
-            return null;
-        }
-        if (values.size() != 1) {
-            throw new AssertionException("expecting header [" + name + "] to be single-valued, but was: " + values);
-        }
-        return values.get(0);
-    }
-
-    @Override
-    public List<String> getHeaderAllValues(String name) {
-        return headers.get(name);
-    }
-
-    @Override
-    public String getBody() {
-        return body;
-    }
-
-    @Override
-    public String wireRepresentation() {
+    public String headerBlock() {
         StringBuilder sb = new StringBuilder(statusLine).append('\n');
         headersInOrder.forEach(header -> sb.append(header.name + ": " + header.value + "\n"));
         if (body != null) {
-            sb.append("\n").append(body);
+            sb.append("\n");
         }
         return sb.toString();
+    }
+
+    @Override
+    public String bodyBlock() {
+        return body;
     }
 
 }